jmt/requirements.txt
Flask==2.2.5
PyJWT==2.7.0
cryptography==40.0.2
